Consolidate the encoding module
