+++ src/derived.ts
@@ -1,5 +1,5 @@
 import { sha256 } from "@noble/hashes/sha2.js";
-import { utf8ToBytes } from "./encoding.js";
+import { utf8ToBytes } from "@noble/hashes/utils.js";
 
 const DETERMINISTIC_PRF_DOMAIN = utf8ToBytes("mera.v1.deterministic.prf");
 // The domain is a compile-time constant, so the salt is constant too: compute
